resolved #109
